+++ enterprise/frontend/src/metabase-enterprise/data_apps/runtime/iframe-vendors.ts
@@ -1,7 +1,7 @@
 import "@mantine/core/styles.css";
 import "@mantine/dates/styles.css";
 
-import "./iframe-baseline.css";
+import "metabase-enterprise/data_apps/sandbox/iframe-baseline.css";
 
 import "metabase/utils/csp";
 import "metabase/utils/dayjs";
