+++ enterprise/frontend/src/metabase-enterprise/data_apps/runtime/lib/describe-error.ts
