+++ enterprise/frontend/src/metabase-enterprise/data_apps/runtime/lib/read-name-from-url.ts
