+++ enterprise/frontend/src/metabase-enterprise/data_apps/runtime/lib/report-error-to-parent.ts
@@ -1,4 +1,4 @@
-import { DATA_APP_ERROR_MESSAGE_TYPE } from "../constants";
+import { DATA_APP_ERROR_MESSAGE_TYPE } from "../../constants";
 
 import type { ErrorDetail } from "./describe-error";
