+++ enterprise/frontend/src/metabase-enterprise/data_apps/runtime/loader.ts
@@ -8,12 +8,11 @@ import { pick } from "underscore";
 import * as sdkExports from "embedding-sdk-package";
 import * as dataAppExports from "embedding-sdk-package/data-app";
 import { getSubpathSafeUrl } from "metabase/urls";
-
+import { createDataAppSandbox } from "metabase-enterprise/data_apps/sandbox/sandbox";
 import {
   DATA_APP_PROVIDER_PROP_KEYS,
   type DataAppMetabaseProviderProps,
-  createDataAppSandbox,
-} from "./sandbox";
+} from "metabase-enterprise/data_apps/sandbox/types";
 
 export interface LoadedDataApp {
   component: React.ComponentType<Record<string, unknown>>;
