+++ frontend/src/embedding-sdk-bundle/lib/create-metabase-query/metric-metadata.ts
