+++ frontend/src/embedding-sdk-bundle/lib/create-metabase-query/validation.ts
