+++ frontend/src/embedding-sdk-bundle/types/ui.ts
@@ -1,4 +1,7 @@
-import type { JSX, ReactNode } from "react";
+export type {
+  SdkErrorComponent,
+  SdkErrorComponentProps,
+} from "metabase/embedding-sdk/types/error-component";
 
 /**
  * @inline
@@ -44,20 +47,6 @@ export type {
   MetabaseColors,
   MetabaseComponentTheme,
 } from "metabase/embedding-sdk/theme";
-
-export type SdkErrorComponentProps = {
-  type?: "relative" | "fixed";
-  message: ReactNode;
-  error?: Error;
-  withCloseButton?: boolean;
-  onClose?: () => void;
-};
-
-export type SdkErrorComponent = ({
-  type,
-  message,
-  error,
-}: SdkErrorComponentProps) => JSX.Element;
 
 export type SdkLoadingError = {
   status: number;
